requirements.txt
PyQt5
PyQt-Fluent-Widgets[full]
Pillow
requests
darkdetect
pyperclip